Move settings UI components
